+++ include/ast/stmt/Stmt.hpp
@@ -12,7 +12,7 @@ namespace AST
         Stmt() = default;
         virtual ~Stmt() = default;
 
-        virtual void print(std::ostream &os, uint8_t tab) const = 0;
+        virtual void print(std::ostream &os, uint32_t tab) const = 0;
     };
 
 } // namespace AST
