+++ include/ast/stmt/types/struct/StructDeclStmt.hpp
@@ -10,7 +10,7 @@ namespace AST
                               DeclarationAccessibility accessibility =
                                   DeclarationAccessibility::Private);
 
-        void print(std::ostream &os, uint8_t tab) const override;
+        void print(std::ostream &os, uint32_t tab) const override;
     };
 
 } // namespace AST
