+++ src/ast/common/IdentifierDecl.cpp
@@ -19,7 +19,7 @@ namespace AST
         return mut_state == IdentifierMutabilityState::Mutable;
     }
 
-    void IdentifierDecl::print(std::ostream &os, uint8_t tab) const
+    void IdentifierDecl::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
