+++ src/ast/stmt/ExprStmt.cpp
@@ -12,7 +12,7 @@ namespace AST
         end_position = this->expr->end_position;
     }
 
-    void ExprStmt::print(std::ostream &os, uint8_t tab) const
+    void ExprStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
