+++ src/ast/stmt/function/FunctionDefStmt.cpp
@@ -16,7 +16,7 @@ namespace AST
         end_position = this->body->end_position;
     }
 
-    void FunctionDefinitionStmt::print(std::ostream &os, uint8_t tab) const
+    void FunctionDefinitionStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
