+++ src/ast/stmt/variable/VariableDeclStmt.cpp
@@ -13,7 +13,7 @@ namespace AST
         end_position = IdentifierDecl::end_position;
     }
 
-    void VariableDeclarationStmt::print(std::ostream &os, uint8_t tab) const
+    void VariableDeclarationStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
