+++ src/ast/stmt/variable/VariableDefStmt.cpp
@@ -19,7 +19,7 @@ namespace AST
             end_position = value->end_position;
     }
 
-    void VariableDefinitionStmt::print(std::ostream &os, uint8_t tab) const
+    void VariableDefinitionStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
